Added a function to produce a ranked list of query results
A function which, when given a name and field of expertise will rank
a set of profiles based on how well they fulfill a combination of both
criteria. The field is looked up in the keywords list, the name is
compared with the query using Levenshtein distance and scaled.
I have left the scaling naively set to 100 for now.